IJPS-73 *wip* remove interface for now.
